Added bodies to the checkPercentageWithinRestraints methods and refactored the class name of DTOMultiLocationQry from DTOMultiLocationQryImpl
