+++ src/test/java/com/sparta/sdets/ConnectionManager.java
@@ -13,10 +13,12 @@ public class ConnectionManager {
     private static final String BASEURL = "https://api.openweathermap.org/";
     private static String endPoint;
 
+    static Properties properties;
+
     public static String getConnection() {
         String inputFile = "src/test/resources/application.properties";
 
-        Properties properties = new Properties();
+        properties = new Properties();
         try {
             properties.load(new java.io.FileReader(inputFile));
         } catch (IOException e) {
