+++ src/test/java/com/sparta/sdets/openweatherdtos/DTOMultiLocationQryOverview.java
@@ -4,5 +4,5 @@ import java.util.ArrayList;
 
 public interface DTOMultiLocationQryOverview {
 
-    boolean checkLocationUniqueness(ArrayList<DTOMultiLocationQryImpl> cities);
+    boolean checkLocationUniqueness(ArrayList<DTOMultiLocationQry> cities);
 }
