+++ src/test/java/com/sparta/sdets/openweatherdtos/DTOMultiLocationQryOverviewImpl.java
@@ -16,7 +16,7 @@ public class DTOMultiLocationQryOverviewImpl implements DTOMultiLocationQryOverv
     private Integer numOfCitiesInBox;
 
     @JsonProperty("list")
-    private ArrayList<DTOMultiLocationQryImpl> cities;
+    private ArrayList<DTOMultiLocationQry> cities;
 
     public DTOMultiLocationQryOverviewImpl(){
 
@@ -46,11 +46,11 @@ public class DTOMultiLocationQryOverviewImpl implements DTOMultiLocationQryOverv
         this.numOfCitiesInBox = numOfCitiesInBox;
     }
 
-    public ArrayList<DTOMultiLocationQryImpl> getCities() {
+    public ArrayList<DTOMultiLocationQry> getCities() {
         return cities;
     }
 
-    public void setCities(ArrayList<DTOMultiLocationQryImpl> cities) {
+    public void setCities(ArrayList<DTOMultiLocationQry> cities) {
         this.cities = cities;
     }
 
@@ -65,7 +65,7 @@ public class DTOMultiLocationQryOverviewImpl implements DTOMultiLocationQryOverv
     }
 
     @Override
-    public boolean checkLocationUniqueness(ArrayList<DTOMultiLocationQryImpl> cities){
+    public boolean checkLocationUniqueness(ArrayList<DTOMultiLocationQry> cities){
         return true;    //TODO
     }
